phase3_llm/requirements.txt
# Phase 3: Groq LLM Integration
groq>=0.4.0
python-dotenv>=1.0.0
